+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## [Unreleased][]
 
+### Changed
+
+-   Abstract scroll and resize event handling, so events could be wrapped with
+    more functionality (e.g. throttle)
+
+### Added
+
+-   `scrollResizeHandler` property
+
+### Removed
+
+-   `debounce` property
+
 ## [1.3.0][] - 2019-07-22
 
 ### Changed
@@ -42,6 +55,6 @@
 
 <!-- prettier-ignore-end -->
 
-
-[Unreleased]: https://github.com/niksy/element-within-viewport/compare/v1.3.0...HEAD
+[unreleased]:
+	https://github.com/niksy/element-within-viewport/compare/v1.3.0...HEAD
 [1.3.0]: https://github.com/niksy/element-within-viewport/tree/v1.3.0